requirements.txt
chrisapp
pypng
numpy
Pillow
tensorflow==1.12.0